Improves code structure by merging some modules
